refactor(queries): split domain hooks and invalidation modules
