+++ src/queries/athlete-profiles.ts
@@ -1,166 +1,4 @@
-import {
-  keepPreviousData,
-  queryOptions,
-  useInfiniteQuery,
-  useMutation,
-  useQuery,
-  useQueryClient,
-} from '@tanstack/react-query';
-import { toast } from 'sonner';
-import type {
-  AthleteProfilesDTO,
-  CheckDuplicateDTO,
-  CreateAthleteProfileDTO,
-  UpdateAthleteProfileDTO,
-} from '@/orpc/athlete-profiles/dto';
-import { DEFAULT_SORTING } from '@/config/athlete';
-import { client } from '@/orpc/client';
-
-export const athleteProfilesDefaultListInput = {
-  page: 1,
-  perPage: 10,
-  filters: [],
-  joinOperator: 'and',
-  sorting: DEFAULT_SORTING,
-} satisfies AthleteProfilesDTO;
-
-export function athleteProfilesQueryOptions(input: AthleteProfilesDTO) {
-  return queryOptions({
-    queryKey: ['athleteProfile', 'list', input] as const,
-    queryFn: () => client.athleteProfile.list(input),
-    staleTime: 30_000,
-    gcTime: 5 * 60 * 1000,
-  });
-}
-
-export function useAthleteProfiles(input: AthleteProfilesDTO) {
-  return useQuery({
-    ...athleteProfilesQueryOptions(input),
-    placeholderData: keepPreviousData,
-  });
-}
-
-export function useAthleteProfilesInfinite(
-  input: Omit<AthleteProfilesDTO, 'page'>
-) {
-  return useInfiniteQuery({
-    queryKey: ['athleteProfile', 'list', 'infinite', input] as const,
-    queryFn: ({ pageParam }: { pageParam: number }) =>
-      client.athleteProfile.list({ ...input, page: pageParam }),
-    initialPageParam: 1,
-    getNextPageParam: (last, allPages) => {
-      const fetched = allPages.reduce((acc, p) => acc + p.items.length, 0);
-      return fetched < last.total ? allPages.length + 1 : undefined;
-    },
-    placeholderData: keepPreviousData,
-  });
-}
-
-function useInvalidateAthleteProfiles() {
-  const queryClient = useQueryClient();
-  return () => {
-    queryClient.invalidateQueries({ queryKey: ['athleteProfile'] });
-    queryClient.invalidateQueries({ queryKey: ['tournamentAthlete'] });
-  };
-}
-
-export function useCreateAthleteProfile(options?: {
-  onSuccess?: () => void;
-  onHardBlock?: () => void;
-  onPossibleDuplicate?: (duplicateIds: Array<string>) => void;
-}) {
-  const invalidate = useInvalidateAthleteProfiles();
-
-  return useMutation({
-    mutationFn: (data: CreateAthleteProfileDTO) =>
-      client.athleteProfile.create(data),
-    onSuccess: () => {
-      invalidate();
-      options?.onSuccess?.();
-    },
-    onError: (err) => {
-      const message = err.message ?? '';
-      if (message.startsWith('DUPLICATE_ATHLETE_CODE_NAME:')) {
-        options?.onHardBlock?.();
-        return;
-      }
-      if (message.startsWith('POSSIBLE_DUPLICATE:')) {
-        try {
-          const ids = JSON.parse(
-            message.replace('POSSIBLE_DUPLICATE: ', '')
-          ) as Array<string>;
-          options?.onPossibleDuplicate?.(ids);
-        } catch {
-          options?.onPossibleDuplicate?.([]);
-        }
-        return;
-      }
-    },
-  });
-}
-
-export function useCheckDuplicate() {
-  return useMutation({
-    mutationFn: (data: CheckDuplicateDTO) =>
-      client.athleteProfile.checkDuplicate(data),
-  });
-}
-
-export function useUpdateAthleteProfile(options?: {
-  onSuccess?: () => void;
-  suppressToast?: boolean;
-}) {
-  const invalidate = useInvalidateAthleteProfiles();
-
-  return useMutation({
-    mutationFn: (data: UpdateAthleteProfileDTO) =>
-      client.athleteProfile.update(data),
-    onSuccess: () => {
-      invalidate();
-      if (!options?.suppressToast) {
-        toast.success('Athlete updated');
-      }
-      options?.onSuccess?.();
-    },
-    onError: (err) => {
-      if (!options?.suppressToast) {
-        toast.error(err.message);
-      }
-    },
-  });
-}
-
-export function useDeleteAthleteProfile(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateAthleteProfiles();
-
-  return useMutation({
-    mutationFn: (data: { id: string }) => client.athleteProfile.delete(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Athlete deleted');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useBulkDeleteAthleteProfiles(options?: {
-  onSuccess?: () => void;
-}) {
-  const invalidate = useInvalidateAthleteProfiles();
-
-  return useMutation({
-    mutationFn: (data: { ids: Array<string> }) =>
-      client.athleteProfile.bulkDelete(data),
-    onSuccess: (deletedCount) => {
-      invalidate();
-      toast.success(
-        deletedCount === 1
-          ? 'Athlete deleted'
-          : `${deletedCount} athletes deleted`
-      );
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
+export * from '@/queries/athlete-profile/athlete-profile-query-options';
+export * from '@/queries/athlete-profile/invalidate-athlete-profile-cache';
+export * from '@/queries/athlete-profile/use-athlete-profile-mutations';
+export * from '@/queries/athlete-profile/use-athlete-profile-queries';
