+++ src/queries/group/use-groups-query.ts
@@ -1,0 +1,6 @@
+import { useQuery } from '@tanstack/react-query';
+import { groupListQueryOptions } from '@/queries/group/group-list-query-options';
+
+export function useGroups(tournamentId: string) {
+  return useQuery(groupListQueryOptions(tournamentId));
+}
