+++ src/queries/groups.ts
@@ -1,159 +1,4 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { toast } from 'sonner';
-import type { QueryClient } from '@tanstack/react-query';
-import type {
-  AssignAthleteDTO,
-  AutoAssignAllDTO,
-  AutoAssignDTO,
-  UnassignAthleteDTO,
-  UpdateGroupDTO,
-} from '@/orpc/groups/dto';
-import { client, orpc } from '@/orpc/client';
-
-export function useGroups(tournamentId: string) {
-  return useQuery(orpc.group.list.queryOptions({ input: { tournamentId } }));
-}
-
-export function isOrpcGroupListQuery(query: {
-  queryKey: ReadonlyArray<unknown>;
-}): boolean {
-  const head = query.queryKey[0];
-  return (
-    Array.isArray(head) &&
-    head.length >= 2 &&
-    head[0] === 'group' &&
-    head[1] === 'list'
-  );
-}
-
-export function invalidateOrpcGroupListQueries(queryClient: QueryClient) {
-  return queryClient.invalidateQueries({ predicate: isOrpcGroupListQuery });
-}
-
-function useInvalidateGroups() {
-  const queryClient = useQueryClient();
-  return () => {
-    void queryClient.invalidateQueries({ queryKey: ['tournament'] });
-    void invalidateOrpcGroupListQueries(queryClient);
-    void queryClient.invalidateQueries({ queryKey: ['tournamentAthlete'] });
-    void queryClient.invalidateQueries({ queryKey: ['activity'] });
-  };
-}
-
-export function useCreateGroup(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: { name: string; tournamentId: string }) =>
-      client.group.create(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Group created');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useUpdateGroup(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: UpdateGroupDTO) => client.group.update(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Group updated');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useDeleteGroup(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: { id: string }) => client.group.delete(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Group deleted');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useAutoAssignGroup(options?: {
-  onSuccess?: () => void;
-  suppressToast?: boolean;
-}) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: AutoAssignDTO) => client.group.autoAssign(data),
-    onSuccess: (result) => {
-      invalidate();
-      if (!options?.suppressToast) {
-        toast.success(`Auto-assigned ${result.assigned} athletes`);
-      }
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useAutoAssignAll(options?: {
-  onSuccess?: (result: {
-    assigned: number;
-    groupsRun: number;
-    groupsSkipped: number;
-  }) => void;
-}) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: AutoAssignAllDTO) => client.group.autoAssignAll(data),
-    onSuccess: (result) => {
-      invalidate();
-      options?.onSuccess?.(result);
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useAssignAthlete(options?: {
-  onSuccess?: () => void;
-  suppressErrorToast?: boolean;
-}) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: AssignAthleteDTO) => client.group.assignAthlete(data),
-    onSuccess: () => {
-      invalidate();
-      options?.onSuccess?.();
-    },
-    onError: (err) => {
-      if (!options?.suppressErrorToast) toast.error(err.message);
-    },
-  });
-}
-
-export function useUnassignAthlete(options?: {
-  onSuccess?: () => void;
-  suppressErrorToast?: boolean;
-}) {
-  const invalidate = useInvalidateGroups();
-
-  return useMutation({
-    mutationFn: (data: UnassignAthleteDTO) =>
-      client.group.unassignAthlete(data),
-    onSuccess: () => {
-      invalidate();
-      options?.onSuccess?.();
-    },
-    onError: (err) => {
-      if (!options?.suppressErrorToast) toast.error(err.message);
-    },
-  });
-}
+export * from '@/queries/group/group-list-query-options';
+export * from '@/queries/group/invalidate-group-cache';
+export * from '@/queries/group/use-group-mutations';
+export * from '@/queries/group/use-groups-query';
