+++ src/queries/matches.ts
@@ -1,349 +1,4 @@
-import {
-  queryOptions,
-  useMutation,
-  useQuery,
-  useQueryClient,
-} from '@tanstack/react-query';
-import { toast } from 'sonner';
-import type {
-  AdminSetMatchStatusDTO,
-  AssignSlotDTO,
-  CreateCustomMatchDTO,
-  GenerateBracketDTO,
-  SetLockDTO,
-  SetWinnerDTO,
-  SwapParticipantsDTO,
-  SwapSlotsDTO,
-  UpdateScoreDTO,
-} from '@/orpc/matches/dto';
-import type { MatchData } from '@/features/dashboard/types';
-import type { BracketDnDMutationContext } from '@/lib/queries/matches';
-import { client } from '@/orpc/client';
-import { invalidateOrpcGroupListQueries } from '@/queries/groups';
-import {
-  applyOptimisticAssign,
-  applyOptimisticSetLock,
-  applyOptimisticSwap,
-  applyOptimisticSwapParticipants,
-  applyOptimisticUpdateScore,
-  findMatchListQueryKey,
-} from '@/lib/queries/matches';
-
-export function useMatches(groupId: string | null) {
-  return useQuery({
-    queryKey: ['match', 'list', groupId] as const,
-    queryFn: () => client.match.list({ groupId: groupId! }),
-    enabled: !!groupId,
-  });
-}
-
-export function tournamentMatchesQueryOptions(tournamentId: string) {
-  return queryOptions({
-    queryKey: ['match', 'list', 'tournament', tournamentId] as const,
-    queryFn: () => client.match.list({ tournamentId }),
-  });
-}
-
-export function useTournamentMatches(tournamentId: string) {
-  return useQuery(tournamentMatchesQueryOptions(tournamentId));
-}
-
-function useInvalidateMatches() {
-  const queryClient = useQueryClient();
-  return () => {
-    void queryClient.invalidateQueries({ queryKey: ['match'] });
-    void queryClient.invalidateQueries({ queryKey: ['tournament'] });
-    void invalidateOrpcGroupListQueries(queryClient);
-    void queryClient.invalidateQueries({ queryKey: ['activity'] });
-    void queryClient.invalidateQueries({
-      predicate: (q) =>
-        Array.isArray(q.queryKey) && q.queryKey[0] === 'advanceSettings',
-    });
-  };
-}
-
-export function useGenerateBracket(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: GenerateBracketDTO) => client.bracket.generate(data),
-    onSuccess: () => {
-      invalidate();
-      options?.onSuccess?.();
-    },
-  });
-}
-
-export function useShuffleBracket(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: { groupId: string }) => client.bracket.shuffle(data),
-    onSuccess: () => {
-      invalidate();
-      options?.onSuccess?.();
-    },
-  });
-}
-
-export function useRegenerateBracket(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: { groupId: string }) => client.bracket.regenerate(data),
-    onSuccess: () => {
-      options?.onSuccess?.();
-    },
-    onSettled: () => {
-      void invalidate();
-    },
-  });
-}
-
-export function useResetBracket(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: { groupId: string }) => client.bracket.reset(data),
-    onSuccess: () => {
-      options?.onSuccess?.();
-    },
-    onSettled: () => {
-      void invalidate();
-    },
-  });
-}
-
-export function useCreateCustomMatch(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: CreateCustomMatchDTO) => client.match.createCustom(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Custom match created');
-      options?.onSuccess?.();
-    },
-    onError: (err) =>
-      toast.error(err instanceof Error ? err.message : 'Request failed'),
-  });
-}
-
-export function useAssignSlot(options?: { onSuccess?: () => void }) {
-  const queryClient = useQueryClient();
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: AssignSlotDTO) => client.match.assignSlot(data),
-    onMutate: async (input): Promise<BracketDnDMutationContext | undefined> => {
-      const queryKey = findMatchListQueryKey(queryClient, input.matchId);
-      if (!queryKey) return undefined;
-
-      await queryClient.cancelQueries({ queryKey });
-      const previousMatches =
-        queryClient.getQueryData<Array<MatchData>>(queryKey);
-      queryClient.setQueryData<Array<MatchData>>(queryKey, (old) => {
-        if (!old) return old;
-        return applyOptimisticAssign(old, input);
-      });
-      return { queryKey, previousMatches };
-    },
-    onError: (_err, _input, context) => {
-      if (context?.queryKey && context.previousMatches !== undefined) {
-        queryClient.setQueryData(context.queryKey, context.previousMatches);
-      }
-    },
-    onSettled: (_data, error) => {
-      void invalidate();
-      if (!error) options?.onSuccess?.();
-    },
-  });
-}
-
-export function useSwapSlots(options?: { onSuccess?: () => void }) {
-  const queryClient = useQueryClient();
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: SwapSlotsDTO) => client.match.swapSlots(data),
-    onMutate: async (input): Promise<BracketDnDMutationContext | undefined> => {
-      const queryKey = findMatchListQueryKey(queryClient, input.matchAId);
-      if (!queryKey) return undefined;
-
-      await queryClient.cancelQueries({ queryKey });
-      const previousMatches =
-        queryClient.getQueryData<Array<MatchData>>(queryKey);
-      queryClient.setQueryData<Array<MatchData>>(queryKey, (old) => {
-        if (!old) return old;
-        return applyOptimisticSwap(old, input);
-      });
-      return { queryKey, previousMatches };
-    },
-    onError: (_err, _input, context) => {
-      if (context?.queryKey && context.previousMatches !== undefined) {
-        queryClient.setQueryData(context.queryKey, context.previousMatches);
-      }
-    },
-    onSettled: (_data, error) => {
-      void invalidate();
-      if (!error) options?.onSuccess?.();
-    },
-  });
-}
-
-export function useSetLock(options?: { onSuccess?: () => void }) {
-  const queryClient = useQueryClient();
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: SetLockDTO) => client.match.setLock(data),
-    onMutate: async (input): Promise<BracketDnDMutationContext | undefined> => {
-      const queryKey = findMatchListQueryKey(queryClient, input.matchId);
-      if (!queryKey) return undefined;
-
-      await queryClient.cancelQueries({ queryKey });
-      const previousMatches =
-        queryClient.getQueryData<Array<MatchData>>(queryKey);
-      queryClient.setQueryData<Array<MatchData>>(queryKey, (old) => {
-        if (!old) return old;
-        return applyOptimisticSetLock(old, input);
-      });
-      return { queryKey, previousMatches };
-    },
-    onError: (err, _input, context) => {
-      if (context?.queryKey && context.previousMatches !== undefined) {
-        queryClient.setQueryData(context.queryKey, context.previousMatches);
-      }
-      toast.error(err.message);
-    },
-    onSettled: (_data, error) => {
-      void invalidate();
-      if (!error) options?.onSuccess?.();
-    },
-  });
-}
-
-export function useUpdateScore(options?: { onSuccess?: () => void }) {
-  const queryClient = useQueryClient();
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: UpdateScoreDTO) => client.match.updateScore(data),
-    onMutate: async (input): Promise<BracketDnDMutationContext | undefined> => {
-      const queryKey = findMatchListQueryKey(queryClient, input.matchId);
-      if (!queryKey) return undefined;
-
-      await queryClient.cancelQueries({ queryKey });
-      const previousMatches =
-        queryClient.getQueryData<Array<MatchData>>(queryKey);
-      queryClient.setQueryData<Array<MatchData>>(queryKey, (old) => {
-        if (!old) return old;
-        return applyOptimisticUpdateScore(old, input);
-      });
-      return { queryKey, previousMatches };
-    },
-    onError: (err, _input, context) => {
-      if (context?.queryKey && context.previousMatches !== undefined) {
-        queryClient.setQueryData(context.queryKey, context.previousMatches);
-      }
-      toast.error(err.message);
-    },
-    onSettled: (_data, error) => {
-      void invalidate();
-      if (!error) {
-        toast.success('Score updated');
-        options?.onSuccess?.();
-      }
-    },
-  });
-}
-
-export function useAdminSetMatchStatus(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: AdminSetMatchStatusDTO) =>
-      client.match.adminSetMatchStatus(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Match status updated');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useResetMatchScore(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: UpdateScoreDTO) => client.match.updateScore(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Match reset');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useDeleteMatch(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: { id: string }) => client.match.delete(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Custom match deleted');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useSetWinner(options?: { onSuccess?: () => void }) {
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: SetWinnerDTO) => client.match.setWinner(data),
-    onSuccess: () => {
-      invalidate();
-      toast.success('Winner set');
-      options?.onSuccess?.();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-}
-
-export function useSwapParticipants(options?: { onSuccess?: () => void }) {
-  const queryClient = useQueryClient();
-  const invalidate = useInvalidateMatches();
-
-  return useMutation({
-    mutationFn: (data: SwapParticipantsDTO) =>
-      client.match.swapParticipants(data),
-    onMutate: async (input): Promise<BracketDnDMutationContext | undefined> => {
-      const queryKey = findMatchListQueryKey(queryClient, input.matchId);
-      if (!queryKey) return undefined;
-
-      await queryClient.cancelQueries({ queryKey });
-      const previousMatches =
-        queryClient.getQueryData<Array<MatchData>>(queryKey);
-      queryClient.setQueryData<Array<MatchData>>(queryKey, (old) => {
-        if (!old) return old;
-        return applyOptimisticSwapParticipants(old, input);
-      });
-      return { queryKey, previousMatches };
-    },
-    onError: (err, _input, context) => {
-      if (context?.queryKey && context.previousMatches !== undefined) {
-        queryClient.setQueryData(context.queryKey, context.previousMatches);
-      }
-      toast.error(err.message);
-    },
-    onSettled: (_data, error) => {
-      void invalidate();
-      if (!error) options?.onSuccess?.();
-    },
-  });
-}
+export * from '@/queries/match/invalidate-match-cache';
+export * from '@/queries/match/match-query-options';
+export * from '@/queries/match/use-match-mutations';
+export * from '@/queries/match/use-match-queries';
