+++ src/queries/tournament-athlete/tournament-athlete-query-options.ts
@@ -1,0 +1,15 @@
+import { queryOptions } from '@tanstack/react-query';
+import type { ListTournamentAthletesDTO } from '@/orpc/tournament-athletes/dto';
+import { tournamentAthleteKeys } from '@/queries/keys';
+import { listTournamentAthletes } from '@/queries/lib/tournament-athlete/tournament-athlete-api';
+
+export function tournamentAthletesQueryOptions(
+  input: ListTournamentAthletesDTO
+) {
+  return queryOptions({
+    queryKey: tournamentAthleteKeys.list(input),
+    queryFn: () => listTournamentAthletes(input),
+  });
+}
+
+export { listAthleteListIdentityKey } from '@/queries/lib/tournament-athlete/list-identity';
