+++ src/routes/dashboard/tournaments/$id.tsx
@@ -1,18 +1,16 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { TournamentPage } from '@/features/dashboard';
 import { TournamentViewerLoading } from '@/features/dashboard/tournament/viewer/loading';
+import { groupListQueryOptions } from '@/queries/group/group-list-query-options';
 import { activityListInfiniteQueryOptions } from '@/queries/activity';
 import { tournamentMatchesQueryOptions } from '@/queries/matches';
 import { tournamentQueryOptions } from '@/queries/tournaments';
-import { orpc } from '@/orpc/client';
 
 export const Route = createFileRoute('/dashboard/tournaments/$id')({
   loader: async ({ params, context: { queryClient } }) => {
     await Promise.all([
       queryClient.ensureQueryData(tournamentQueryOptions(params.id)),
-      queryClient.ensureQueryData(
-        orpc.group.list.queryOptions({ input: { tournamentId: params.id } })
-      ),
+      queryClient.ensureQueryData(groupListQueryOptions(params.id)),
       queryClient.ensureQueryData(tournamentMatchesQueryOptions(params.id)),
     ]);
     await queryClient.prefetchInfiniteQuery({
